+++ BE/sellerB/src/main/java/backend/sellerB/repository/ManagerRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ManagerRepository extends JpaRepository<Manager, Integer> {
+public interface ManagerRepository extends JpaRepository<Manager, Long> {
     @EntityGraph(attributePaths = "authorities") // EntityGraph annotation은 쿼리가 수행될때 Lazy조회가 아니고 Eager조회로 authorities 정보를 같이 가져옴
     Optional<Manager> findBymanagerId(String managerId);
 
-    Optional<Manager> findByManagerSeq(Integer Seq);
+    Optional<Manager> findByManagerSeq(Long Seq);
 }
